Added single target instant heal and regenerative heal, player movement is now locked while casting, and magic actually heals

+++ HealingOverflow/Assets/Scripts/SelfAreaHeal.cs
@@ -3,7 +3,7 @@ public class SelfAreaHeal : Magic
 {
 	[SerializeField] private GameObject healingCirclePrefab;
     [SerializeField] private float healingCircleDuration;
-    [SerializeField] private float healingCirclePower;
+    [SerializeField] private int healingCirclePower;
     protected override void OnCast()
     {
         HealingCircle healingCircle = Instantiate(healingCirclePrefab, transform.position, transform.rotation).GetComponent<HealingCircle>();
